Added coordinates needed by the model
Starting with alt/azi coordinates, we need an extra flip
of the x axis in order to get the batoid/model coordinate system.
I add extra columns with prefix 'oc' to the table. Note that the
field coordinates need to be based on these oc coordinates. I
tested the resulting moments table against the model.

	modified:   PSF RA Plot_imsim.ipynb
	modified:   ellipticity_util.py